+++ src/sketch.h
@@ -8,260 +8,21 @@
 #ifndef SOLVESPACE_SKETCH_H
 #define SOLVESPACE_SKETCH_H
 
-#include <ctype.h>
-#include <stdint.h>
-#include "dsc.h"
-#include "platform.h"
-#include "list.h"
 #include "smesh.h"
 #include "sedge.h"
 #include "group.h"
-#include "handle.h"
 #include "param.h"
 #include "entity.h"
+#include "request.h"
+#include "equation.h"
+#include "render/render.h"
+#include "platform/gui.h"
+#include "constraint.h"
+#include "style.h"
 
 namespace SolveSpace {
 
-enum class PolyError : uint32_t {
-    GOOD              = 0,
-    NOT_CLOSED        = 1,
-    NOT_COPLANAR      = 2,
-    SELF_INTERSECTING = 3,
-    ZERO_LEN_EDGE     = 4
-};
-
-enum class StipplePattern : uint32_t {
-    CONTINUOUS     = 0,
-    SHORT_DASH     = 1,
-    DASH           = 2,
-    LONG_DASH      = 3,
-    DASH_DOT       = 4,
-    DASH_DOT_DOT   = 5,
-    DOT            = 6,
-    FREEHAND       = 7,
-    ZIGZAG         = 8,
-
-    LAST           = ZIGZAG
-};
-
-const std::vector<double> &StipplePatternDashes(StipplePattern pattern);
-double StipplePatternLength(StipplePattern pattern);
-
 enum class Command : uint32_t;
-
-
-class EntityId {
-public:
-    uint32_t v;     // entity ID, starting from 0
-};
-class EntityMap {
-public:
-    int         tag;
-
-    EntityId    h;
-    hEntity     input;
-    int         copyNumber;
-    // (input, copyNumber) gets mapped to ((Request)xxx).entity(h.v)
-
-    void Clear() {}
-};
-
-// A user request for some primitive or derived operation; for example a
-// line, or a step and repeat.
-class Request {
-public:
-    // Some predefined requests, that are present in every sketch.
-    static const hRequest   HREQUEST_REFERENCE_XY;
-    static const hRequest   HREQUEST_REFERENCE_YZ;
-    static const hRequest   HREQUEST_REFERENCE_ZX;
-
-    int         tag;
-    hRequest    h;
-
-    // Types of requests
-    enum class Type : uint32_t {
-        WORKPLANE              = 100,
-        DATUM_POINT            = 101,
-        LINE_SEGMENT           = 200,
-        CUBIC                  = 300,
-        CUBIC_PERIODIC         = 301,
-        CIRCLE                 = 400,
-        ARC_OF_CIRCLE          = 500,
-        TTF_TEXT               = 600,
-        IMAGE                  = 700
-    };
-
-    Request::Type type;
-    int         extraPoints;
-
-    hEntity     workplane; // or Entity::FREE_IN_3D
-    hGroup      group;
-    hStyle      style;
-
-    bool        construction;
-
-    std::string str;
-    std::string font;
-    Platform::Path file;
-    double      aspectRatio;
-
-    static hParam AddParam(ParamList *param, hParam hp);
-    void Generate(EntityList *entity, ParamList *param);
-
-    std::string DescriptionString() const;
-    int IndexOfPoint(hEntity he) const;
-
-    void Clear() {}
-};
-
-#define MAX_POINTS_IN_ENTITY (12)
-class EntityBase {
-public:
-    int         tag;
-    hEntity     h;
-
-    static const hEntity    FREE_IN_3D;
-    static const hEntity    NO_ENTITY;
-
-    enum class Type : uint32_t {
-        POINT_IN_3D            =  2000,
-        POINT_IN_2D            =  2001,
-        POINT_N_TRANS          =  2010,
-        POINT_N_ROT_TRANS      =  2011,
-        POINT_N_COPY           =  2012,
-        POINT_N_ROT_AA         =  2013,
-
-        NORMAL_IN_3D           =  3000,
-        NORMAL_IN_2D           =  3001,
-        NORMAL_N_COPY          =  3010,
-        NORMAL_N_ROT           =  3011,
-        NORMAL_N_ROT_AA        =  3012,
-
-        DISTANCE               =  4000,
-        DISTANCE_N_COPY        =  4001,
-
-        FACE_NORMAL_PT         =  5000,
-        FACE_XPROD             =  5001,
-        FACE_N_ROT_TRANS       =  5002,
-        FACE_N_TRANS           =  5003,
-        FACE_N_ROT_AA          =  5004,
-
-        WORKPLANE              = 10000,
-        LINE_SEGMENT           = 11000,
-        CUBIC                  = 12000,
-        CUBIC_PERIODIC         = 12001,
-        CIRCLE                 = 13000,
-        ARC_OF_CIRCLE          = 14000,
-        TTF_TEXT               = 15000,
-        IMAGE                  = 16000
-    };
-
-    Type        type;
-
-    hGroup      group;
-    hEntity     workplane;   // or Entity::FREE_IN_3D
-
-    // When it comes time to draw an entity, we look here to get the
-    // defining variables.
-    hEntity     point[MAX_POINTS_IN_ENTITY];
-    int         extraPoints;
-    hEntity     normal;
-    hEntity     distance;
-    // The only types that have their own params are points, normals,
-    // and directions.
-    hParam      param[7];
-
-    // Transformed points/normals/distances have their numerical base
-    Vector      numPoint;
-    Quaternion  numNormal;
-    double      numDistance;
-
-    std::string str;
-    std::string font;
-    Platform::Path file;
-    double      aspectRatio;
-
-    // For entities that are derived by a transformation, the number of
-    // times to apply the transformation.
-    int timesApplied;
-
-    Quaternion GetAxisAngleQuaternion(int param0) const;
-    ExprQuaternion GetAxisAngleQuaternionExprs(int param0) const;
-
-    bool IsCircle() const;
-    Expr *CircleGetRadiusExpr() const;
-    double CircleGetRadiusNum() const;
-    void ArcGetAngles(double *thetaa, double *thetab, double *dtheta) const;
-
-    bool HasVector() const;
-    ExprVector VectorGetExprs() const;
-    ExprVector VectorGetExprsInWorkplane(hEntity wrkpl) const;
-    Vector VectorGetNum() const;
-    Vector VectorGetRefPoint() const;
-    Vector VectorGetStartPoint() const;
-
-    // For distances
-    bool IsDistance() const;
-    double DistanceGetNum() const;
-    Expr *DistanceGetExpr() const;
-    void DistanceForceTo(double v);
-
-    bool IsWorkplane() const;
-    // The plane is points P such that P dot (xn, yn, zn) - d = 0
-    void WorkplaneGetPlaneExprs(ExprVector *n, Expr **d) const;
-    ExprVector WorkplaneGetOffsetExprs() const;
-    Vector WorkplaneGetOffset() const;
-    EntityBase *Normal() const;
-
-    bool IsFace() const;
-    ExprVector FaceGetNormalExprs() const;
-    Vector FaceGetNormalNum() const;
-    ExprVector FaceGetPointExprs() const;
-    Vector FaceGetPointNum() const;
-
-    bool IsPoint() const;
-    // Applies for any of the point types
-    Vector PointGetNum() const;
-    ExprVector PointGetExprs() const;
-    void PointGetExprsInWorkplane(hEntity wrkpl, Expr **u, Expr **v) const;
-    ExprVector PointGetExprsInWorkplane(hEntity wrkpl) const;
-    void PointForceTo(Vector v);
-    void PointForceParamTo(Vector v);
-    // These apply only the POINT_N_ROT_TRANS, which has an assoc rotation
-    Quaternion PointGetQuaternion() const;
-    void PointForceQuaternionTo(Quaternion q);
-
-    bool IsNormal() const;
-    // Applies for any of the normal types
-    Quaternion NormalGetNum() const;
-    ExprQuaternion NormalGetExprs() const;
-    void NormalForceTo(Quaternion q);
-
-    Vector NormalU() const;
-    Vector NormalV() const;
-    Vector NormalN() const;
-    ExprVector NormalExprsU() const;
-    ExprVector NormalExprsV() const;
-    ExprVector NormalExprsN() const;
-
-    Vector CubicGetStartNum() const;
-    Vector CubicGetFinishNum() const;
-    ExprVector CubicGetStartTangentExprs() const;
-    ExprVector CubicGetFinishTangentExprs() const;
-    Vector CubicGetStartTangentNum() const;
-    Vector CubicGetFinishTangentNum() const;
-
-    bool HasEndpoints() const;
-    Vector EndpointStart() const;
-    Vector EndpointFinish() const;
-
-    void RectGetPointsExprs(ExprVector *eap, ExprVector *ebp) const;
-
-    void AddEq(IdList<Equation,hEquation> *l, Expr *expr, int index) const;
-    void GenerateEquations(IdList<Equation,hEquation> *l) const;
-
-    void Clear() {}
-};
 
 class EntReqTable {
 public:
@@ -272,292 +33,41 @@ public:
     static Request::Type GetRequestForEntity(EntityBase::Type ent);
 };
 
-
-class ConstraintBase {
+#ifdef LIBRARY
+#   define ENTITY EntityBase
+#   define CONSTRAINT ConstraintBase
+#else
+#   define ENTITY Entity
+#   define CONSTRAINT Constraint
+#endif
+class Sketch {
 public:
-    int         tag;
-    hConstraint h;
+    // These are user-editable, and define the sketch.
+    IdList<Group,hGroup>            group;
+    List<hGroup>                    groupOrder;
+    IdList<CONSTRAINT,hConstraint>  constraint;
+    IdList<Request,hRequest>        request;
+    IdList<Style,hStyle>            style;
 
-    static const hConstraint NO_CONSTRAINT;
+    // These are generated from the above.
+    IdList<ENTITY,hEntity>          entity;
+    IdList<Param,hParam>            param;
 
-    enum class Type : uint32_t {
-        POINTS_COINCIDENT      =  20,
-        PT_PT_DISTANCE         =  30,
-        PT_PLANE_DISTANCE      =  31,
-        PT_LINE_DISTANCE       =  32,
-        PT_FACE_DISTANCE       =  33,
-        PROJ_PT_DISTANCE       =  34,
-        PT_IN_PLANE            =  41,
-        PT_ON_LINE             =  42,
-        PT_ON_FACE             =  43,
-        EQUAL_LENGTH_LINES     =  50,
-        LENGTH_RATIO           =  51,
-        EQ_LEN_PT_LINE_D       =  52,
-        EQ_PT_LN_DISTANCES     =  53,
-        EQUAL_ANGLE            =  54,
-        EQUAL_LINE_ARC_LEN     =  55,
-        LENGTH_DIFFERENCE      =  56,
-        SYMMETRIC              =  60,
-        SYMMETRIC_HORIZ        =  61,
-        SYMMETRIC_VERT         =  62,
-        SYMMETRIC_LINE         =  63,
-        AT_MIDPOINT            =  70,
-        HORIZONTAL             =  80,
-        VERTICAL               =  81,
-        DIAMETER               =  90,
-        PT_ON_CIRCLE           = 100,
-        SAME_ORIENTATION       = 110,
-        ANGLE                  = 120,
-        PARALLEL               = 121,
-        PERPENDICULAR          = 122,
-        ARC_LINE_TANGENT       = 123,
-        CUBIC_LINE_TANGENT     = 124,
-        CURVE_CURVE_TANGENT    = 125,
-        EQUAL_RADIUS           = 130,
-        WHERE_DRAGGED          = 200,
+    inline CONSTRAINT *GetConstraint(hConstraint h)
+    { return constraint.FindById(h); }
+    inline ENTITY  *GetEntity (hEntity  h) { return entity. FindById(h); }
+    inline Param   *GetParam  (hParam   h) { return param.  FindById(h); }
+    inline Request *GetRequest(hRequest h) { return request.FindById(h); }
+    inline Group   *GetGroup  (hGroup   h) { return group.  FindById(h); }
+    // Styles are handled a bit differently.
 
-        COMMENT                = 1000
-    };
+    void Clear();
 
-    Type        type;
-
-    hGroup      group;
-    hEntity     workplane;
-
-    // These are the parameters for the constraint.
-    double      valA;
-    hParam      valP;
-    hEntity     ptA;
-    hEntity     ptB;
-    hEntity     entityA;
-    hEntity     entityB;
-    hEntity     entityC;
-    hEntity     entityD;
-    bool        other;
-    bool        other2;
-
-    bool        reference;  // a ref dimension, that generates no eqs
-    std::string comment;    // since comments are represented as constraints
-
-    bool HasLabel() const;
-
-    void Generate(IdList<Param, hParam> *param);
-
-    void GenerateEquations(IdList<Equation,hEquation> *entity,
-                           bool forReference = false) const;
-    // Some helpers when generating symbolic constraint equations
-    void ModifyToSatisfy();
-    void AddEq(IdList<Equation,hEquation> *l, Expr *expr, int index) const;
-    void AddEq(IdList<Equation,hEquation> *l, const ExprVector &v, int baseIndex = 0) const;
-    static Expr *DirectionCosine(hEntity wrkpl, ExprVector ae, ExprVector be);
-    static Expr *Distance(hEntity workplane, hEntity pa, hEntity pb);
-    static Expr *PointLineDistance(hEntity workplane, hEntity pt, hEntity ln);
-    static Expr *PointPlaneDistance(ExprVector p, hEntity plane);
-    static ExprVector VectorsParallel3d(ExprVector a, ExprVector b, hParam p);
-    static ExprVector PointInThreeSpace(hEntity workplane, Expr *u, Expr *v);
-
-    void Clear() {}
+    BBox CalculateEntityBBox(bool includingInvisible);
+    Group *GetRunningMeshGroupFor(hGroup h);
 };
-
-class Constraint : public ConstraintBase {
-public:
-    // See Entity::Entity().
-    Constraint() : ConstraintBase({}), disp() {}
-
-    // These define how the constraint is drawn on-screen.
-    struct {
-        Vector      offset;
-        hStyle      style;
-    } disp;
-
-    bool IsVisible() const;
-    bool IsStylable() const;
-    hStyle GetStyle() const;
-    bool HasLabel() const;
-    std::string Label() const;
-
-    enum class DrawAs { DEFAULT, HOVERED, SELECTED };
-    void Draw(DrawAs how, Canvas *canvas);
-    Vector GetLabelPos(const Camera &camera);
-    void GetReferencePoints(const Camera &camera, std::vector<Vector> *refs);
-
-    void DoLayout(DrawAs how, Canvas *canvas,
-                  Vector *labelPos, std::vector<Vector> *refs);
-    void DoLine(Canvas *canvas, Canvas::hStroke hcs, Vector a, Vector b);
-    void DoStippledLine(Canvas *canvas, Canvas::hStroke hcs, Vector a, Vector b);
-    bool DoLineExtend(Canvas *canvas, Canvas::hStroke hcs,
-                      Vector p0, Vector p1, Vector pt, double salient);
-    void DoArcForAngle(Canvas *canvas, Canvas::hStroke hcs,
-                       Vector a0, Vector da, Vector b0, Vector db,
-                       Vector offset, Vector *ref, bool trim);
-    void DoArrow(Canvas *canvas, Canvas::hStroke hcs,
-                 Vector p, Vector dir, Vector n, double width, double angle, double da);
-    void DoLineWithArrows(Canvas *canvas, Canvas::hStroke hcs,
-                          Vector ref, Vector a, Vector b, bool onlyOneExt);
-    int  DoLineTrimmedAgainstBox(Canvas *canvas, Canvas::hStroke hcs,
-                                 Vector ref, Vector a, Vector b, bool extend,
-                                 Vector gr, Vector gu, double swidth, double sheight);
-    int  DoLineTrimmedAgainstBox(Canvas *canvas, Canvas::hStroke hcs,
-                                 Vector ref, Vector a, Vector b, bool extend = true);
-    void DoLabel(Canvas *canvas, Canvas::hStroke hcs,
-                 Vector ref, Vector *labelPos, Vector gr, Vector gu);
-    void DoProjectedPoint(Canvas *canvas, Canvas::hStroke hcs, Vector *p);
-    void DoProjectedPoint(Canvas *canvas, Canvas::hStroke hcs, Vector *p, Vector n, Vector o);
-
-    void DoEqualLenTicks(Canvas *canvas, Canvas::hStroke hcs,
-                         Vector a, Vector b, Vector gn, Vector *refp);
-    void DoEqualRadiusTicks(Canvas *canvas, Canvas::hStroke hcs,
-                            hEntity he, Vector *refp);
-
-    std::string DescriptionString() const;
-
-    static hConstraint AddConstraint(Constraint *c, bool rememberForUndo);
-    static hConstraint AddConstraint(Constraint *c);
-    static void MenuConstrain(Command id);
-    static void DeleteAllConstraintsFor(Constraint::Type type, hEntity entityA, hEntity ptA);
-
-    static hConstraint ConstrainCoincident(hEntity ptA, hEntity ptB);
-    static hConstraint Constrain(Constraint::Type type, hEntity ptA, hEntity ptB, hEntity entityA);
-    static hConstraint Constrain(Constraint::Type type, hEntity ptA, hEntity ptB,
-                                    hEntity entityA, hEntity entityB,
-                                    bool other, bool other2);
-};
-
-class Equation {
-public:
-    int         tag;
-    hEquation   h;
-
-    Expr        *e;
-
-    void Clear() {}
-};
-
-class Style {
-public:
-    int         tag;
-    hStyle      h;
-
-    enum {
-        // If an entity has no style, then it will be colored according to
-        // whether the group that it's in is active or not, whether it's
-        // construction or not, and so on.
-        NO_STYLE       = 0,
-
-        ACTIVE_GRP     = 1,
-        CONSTRUCTION   = 2,
-        INACTIVE_GRP   = 3,
-        DATUM          = 4,
-        SOLID_EDGE     = 5,
-        CONSTRAINT     = 6,
-        SELECTED       = 7,
-        HOVERED        = 8,
-        CONTOUR_FILL   = 9,
-        NORMALS        = 10,
-        ANALYZE        = 11,
-        DRAW_ERROR     = 12,
-        DIM_SOLID      = 13,
-        HIDDEN_EDGE    = 14,
-        OUTLINE        = 15,
-
-        FIRST_CUSTOM   = 0x100
-    };
-
-    std::string name;
-
-    enum class UnitsAs : uint32_t {
-        PIXELS   = 0,
-        MM       = 1
-    };
-    double      width;
-    UnitsAs     widthAs;
-    double      textHeight;
-    UnitsAs     textHeightAs;
-    enum class TextOrigin : uint32_t {
-        NONE    = 0x00,
-        LEFT    = 0x01,
-        RIGHT   = 0x02,
-        BOT     = 0x04,
-        TOP     = 0x08
-    };
-    TextOrigin  textOrigin;
-    double      textAngle;
-    RgbaColor   color;
-    bool        filled;
-    RgbaColor   fillColor;
-    bool        visible;
-    bool        exportable;
-    StipplePattern stippleType;
-    double      stippleScale;
-    int         zIndex;
-
-    // The default styles, for entities that don't have a style assigned yet,
-    // and for datums and such.
-    typedef struct {
-        hStyle      h;
-        const char *cnfPrefix;
-        RgbaColor   color;
-        double      width;
-        int         zIndex;
-    } Default;
-    static const Default Defaults[];
-
-    static std::string CnfColor(const std::string &prefix);
-    static std::string CnfWidth(const std::string &prefix);
-    static std::string CnfTextHeight(const std::string &prefix);
-    static std::string CnfPrefixToName(const std::string &prefix);
-
-    static void CreateAllDefaultStyles();
-    static void CreateDefaultStyle(hStyle h);
-    static void FillDefaultStyle(Style *s, const Default *d = NULL, bool factory = false);
-    static void FreezeDefaultStyles(Platform::SettingsRef settings);
-    static void LoadFactoryDefaults();
-
-    static void AssignSelectionToStyle(uint32_t v);
-    static uint32_t CreateCustomStyle(bool rememberForUndo = true);
-
-    static RgbaColor RewriteColor(RgbaColor rgb);
-
-    static Style *Get(hStyle hs);
-    static RgbaColor Color(hStyle hs, bool forExport=false);
-    static RgbaColor Color(int hs, bool forExport=false);
-    static RgbaColor FillColor(hStyle hs, bool forExport=false);
-    static double Width(hStyle hs);
-    static double Width(int hs);
-    static double WidthMm(int hs);
-    static double TextHeight(hStyle hs);
-    static double DefaultTextHeight();
-    static Canvas::Stroke Stroke(hStyle hs);
-    static Canvas::Stroke Stroke(int hs);
-    static bool Exportable(int hs);
-    static hStyle ForEntity(hEntity he);
-    static StipplePattern PatternType(hStyle hs);
-    static double StippleScaleMm(hStyle hs);
-
-    std::string DescriptionString() const;
-
-    void Clear() {}
-};
-
-
-// The format for entities stored on the clipboard.
-class ClipboardRequest {
-public:
-    Request::Type type;
-    int         extraPoints;
-    hStyle      style;
-    std::string str;
-    std::string font;
-    Platform::Path file;
-    bool        construction;
-
-    Vector      point[MAX_POINTS_IN_ENTITY];
-    double      distance;
-
-    hEntity     oldEnt;
-    hEntity     oldPointEnt[MAX_POINTS_IN_ENTITY];
-    hRequest    newReq;
-};
+#undef ENTITY
+#undef CONSTRAINT
 
 }
