+++ src/solvespaceui.h
@@ -14,6 +14,8 @@
 #include "resource.h"
 #include "list.h"
 #include "system.h"
+#include "filewriters.h"
+#include "ttf.h"
 
 namespace SolveSpace {
 
@@ -21,6 +23,25 @@ enum class Unit : uint32_t {
     MM = 0,
     INCHES,
     METERS
+};
+
+// The format for entities stored on the clipboard.
+class ClipboardRequest {
+public:
+    Request::Type type;
+    int         extraPoints;
+    hStyle      style;
+    std::string str;
+    std::string font;
+    Platform::Path file;
+    bool        construction;
+
+    Vector      point[MAX_POINTS_IN_ENTITY];
+    double      distance;
+
+    hEntity     oldEnt;
+    hEntity     oldPointEnt[MAX_POINTS_IN_ENTITY];
+    hRequest    newReq;
 };
 
 class SolveSpaceUI {
