+++ src/ttf.h
@@ -9,6 +9,19 @@
 #ifndef SOLVESPACE_TTF_H
 #define SOLVESPACE_TTF_H
 
+#include <string>
+#include "platform.h"
+#include "polygon.h"
+
+// We declare these in advance instead of simply using FT_Library
+// (defined as typedef FT_LibraryRec_* FT_Library) because including
+// freetype.h invokes indescribable horrors and we would like to avoid
+// doing that every time we include solvespace.h.
+struct FT_LibraryRec_;
+struct FT_FaceRec_;
+
+namespace SolveSpace {
+
 class TtfFont {
 public:
     Platform::Path  fontFile;
@@ -40,5 +53,7 @@ public:
                     SBezierList *sbl, Vector origin, Vector u, Vector v);
     double AspectRatio(const std::string &font, const std::string &str);
 };
+
+}
 
 #endif
